feat: simply fuse daemon and runtime commands without any modification
